requirements.txt
Flask
PyPDF2
gunicorn
numpy
python-dotenv
requests
scikit-learn
sentence-transformers


